fix(habitos): mover useSearchParams a componente cliente envuelto en Suspense
